moved tests out of the src directory
