+++ .gitignore
@@ -12,17 +12,17 @@ tools/tspawn/defaults
 *.moc
 
 
-src/test/buildtest/buildtest
-src/test/fieldnametovariablename/fieldnametovariablename
-src/test/hmac/hmac
-src/test/htmlescape/htmlescape
-src/test/htmlparser/htmlparser
-src/test/httpheader/httpheader
-src/test/mailmessage/mailmessage
-src/test/multipartformdata/multipartformdata
-src/test/paginator/paginator
-src/test/rand/rand
-src/test/sharedmemorylogstream/sharedmemorylogstream
-src/test/smtpmailer/smtpmailer
-src/test/viewhelper/viewhelper
-src/test/urlrouter/urlrouter
+/test/buildtest/buildtest
+/test/fieldnametovariablename/fieldnametovariablename
+/test/hmac/hmac
+/test/htmlescape/htmlescape
+/test/htmlparser/htmlparser
+/test/httpheader/httpheader
+/test/mailmessage/mailmessage
+/test/multipartformdata/multipartformdata
+/test/paginator/paginator
+/test/rand/rand
+/test/sharedmemorylogstream/sharedmemorylogstream
+/test/smtpmailer/smtpmailer
+/test/viewhelper/viewhelper
+/test/urlrouter/urlrouter
