--- src/test/test.pro
@@ -1,15 +1,0 @@
-TEMPLATE=subdirs
-SUBDIRS += buildtest
-SUBDIRS += fieldnametovariablename
-SUBDIRS += hmac
-SUBDIRS += htmlescape
-SUBDIRS += htmlparser
-SUBDIRS += httpheader
-SUBDIRS += mailmessage
-SUBDIRS += multipartformdata
-SUBDIRS += paginator
-SUBDIRS += rand
-SUBDIRS += sharedmemorylogstream
-SUBDIRS += smtpmailer
-SUBDIRS += urlrouter
-SUBDIRS += viewhelper
