+++ test/buildtest/blogobject.h
