+++ test/buildtest/fooobject.h
