+++ test/fieldnametovariablename/fieldnametovariablename.pro
@@ -1,0 +1,8 @@
+include(../test.pri)
+
+TARGET=fieldnametovariablename
+QT += network
+
+SOURCES = main.cpp
+
+
