+++ test/hmac/hmac.pro
@@ -1,0 +1,6 @@
+include(../test.pri)
+
+TARGET = hmac
+QT += network
+
+SOURCES = main.cpp
