+++ test/hmac/main.cpp
