+++ test/htmlescape/htmlescape.pro
@@ -1,0 +1,5 @@
+include(../test.pri)
+
+TARGET = htmlescape
+
+SOURCES = main.cpp
