+++ test/htmlparser/htmlparser.pro
@@ -1,0 +1,5 @@
+include(../test.pri)
+
+TARGET = htmlparser
+
+SOURCES = htmlparser.cpp
