+++ test/httpheader/httpheader.pro
@@ -1,0 +1,6 @@
+include(../test.pri)
+
+TARGET = httpheader
+QT += network
+
+SOURCES = main.cpp
