+++ test/mailmessage/mailmessage.pro
@@ -1,0 +1,5 @@
+include(../test.pri)
+
+TARGET = mailmessage
+
+SOURCES = main.cpp
