+++ test/multipartformdata/multipartformdata.pro
@@ -1,0 +1,6 @@
+include(../test.pri)
+
+TARGET = multipartformdata
+QT += network sql
+
+SOURCES = multipartformdata.cpp
