+++ test/paginator/paginator.pro
@@ -1,0 +1,6 @@
+include(../test.pri)
+
+TARGET = paginator
+QT += network sql
+
+SOURCES = paginator.cpp
