+++ test/rand/rand.pro
@@ -1,0 +1,6 @@
+include(../test.pri)
+
+TARGET = rand
+QT += network
+
+SOURCES = main.cpp
