+++ test/sharedmemorylogstream/sharedmemorylogstream.pro
@@ -1,0 +1,6 @@
+include(../test.pri)
+
+TARGET = sharedmemorylogstream
+QT += network sql
+
+SOURCES += benchmarking.cpp
