+++ test/smtpmailer/smtpmailer.pro
@@ -1,0 +1,6 @@
+include(../test.pri)
+
+TARGET = smtpmailer
+QT += network
+
+SOURCES = main.cpp
