+++ test/test.pro
@@ -1,0 +1,17 @@
+TEMPLATE = subdirs
+CONFIG += ordered
+
+SUBDIRS += buildtest
+SUBDIRS += fieldnametovariablename
+SUBDIRS += hmac
+SUBDIRS += htmlescape
+SUBDIRS += htmlparser
+SUBDIRS += httpheader
+SUBDIRS += mailmessage
+SUBDIRS += multipartformdata
+SUBDIRS += paginator
+SUBDIRS += rand
+SUBDIRS += sharedmemorylogstream
+SUBDIRS += smtpmailer
+SUBDIRS += urlrouter
+SUBDIRS += viewhelper
