+++ test/urlrouter/urlrouter.pro
@@ -1,0 +1,5 @@
+include(../test.pri)
+
+TARGET = urlrouter
+
+SOURCES = main.cpp
