+++ test/viewhelper/viewhelper.pro
@@ -1,0 +1,6 @@
+include(../test.pri)
+
+TARGET = viewhelper
+QT += network sql
+
+SOURCES = viewhelper.cpp
